Test migration path generation
Part of #32738

There are no functional project changes here, Instead this PR adds code to allow migrations to generated without actually creating the files. This allows testing of the current behavior such that we can make changes and have confidence that we know what the impact is.

+++ src/EFCore.Design/Migrations/Design/MigrationsScaffolder.cs
@@ -61,12 +61,14 @@ public class MigrationsScaffolder : IMigrationsScaffolder
     ///     the sub-namespace should not both be empty.
     /// </param>
     /// <param name="language">The project's language.</param>
+    /// <param name="dryRun">If true, then nothing is actually written to disk.</param>
     /// <returns>The scaffolded migration.</returns>
     public virtual ScaffoldedMigration ScaffoldMigration(
         string migrationName,
         string? rootNamespace,
         string? subNamespace = null,
-        string? language = null)
+        string? language = null,
+        bool dryRun = false)
     {
         if (string.Equals(migrationName, "migration", StringComparison.OrdinalIgnoreCase))
         {
@@ -222,9 +224,10 @@ public class MigrationsScaffolder : IMigrationsScaffolder
     /// <param name="projectDir">The project's root directory.</param>
     /// <param name="rootNamespace">The project's root namespace.</param>
     /// <param name="force">Don't check to see if the migration has been applied to the database.</param>
+    /// <param name="dryRun">If true, then nothing is actually written to disk.</param>
     /// <returns>The removed migration files.</returns>
-    public virtual MigrationFiles RemoveMigration(string projectDir, string rootNamespace, bool force)
-        => RemoveMigration(projectDir, rootNamespace, force, language: null);
+    public virtual MigrationFiles RemoveMigration(string projectDir, string rootNamespace, bool force, bool dryRun)
+        => RemoveMigration(projectDir, rootNamespace, force, language: null, dryRun: false);
 
     /// <summary>
     ///     Removes the previous migration.
@@ -233,13 +236,15 @@ public class MigrationsScaffolder : IMigrationsScaffolder
     /// <param name="rootNamespace">The project's root namespace.</param>
     /// <param name="force">Don't check to see if the migration has been applied to the database.</param>
     /// <param name="language">The project's language.</param>
+    /// <param name="dryRun">If true, then nothing is actually written to disk.</param>
     /// <returns>The removed migration files.</returns>
     // TODO: DRY (file names)
     public virtual MigrationFiles RemoveMigration(
         string projectDir,
         string? rootNamespace,
         bool force,
-        string? language)
+        string? language,
+        bool dryRun)
     {
         var files = new MigrationFiles();
 
@@ -296,7 +301,11 @@ public class MigrationsScaffolder : IMigrationsScaffolder
                 if (migrationFile != null)
                 {
                     Dependencies.OperationReporter.WriteInformation(DesignStrings.RemovingMigration(migration.GetId()));
-                    File.Delete(migrationFile);
+                    if (!dryRun)
+                    {
+                        File.Delete(migrationFile);
+                    }
+
                     files.MigrationFile = migrationFile;
                 }
                 else
@@ -309,7 +318,11 @@ public class MigrationsScaffolder : IMigrationsScaffolder
                 var migrationMetadataFile = TryGetProjectFile(projectDir, migrationMetadataFileName);
                 if (migrationMetadataFile != null)
                 {
-                    File.Delete(migrationMetadataFile);
+                    if (!dryRun)
+                    {
+                        File.Delete(migrationMetadataFile);
+                    }
+
                     files.MetadataFile = migrationMetadataFile;
                 }
                 else
@@ -336,7 +349,11 @@ public class MigrationsScaffolder : IMigrationsScaffolder
             if (modelSnapshotFile != null)
             {
                 Dependencies.OperationReporter.WriteInformation(DesignStrings.RemovingSnapshot);
-                File.Delete(modelSnapshotFile);
+                if (!dryRun)
+                {
+                    File.Delete(modelSnapshotFile);
+                }
+
                 files.SnapshotFile = modelSnapshotFile;
             }
             else
@@ -363,7 +380,11 @@ public class MigrationsScaffolder : IMigrationsScaffolder
                 modelSnapshotFileName);
 
             Dependencies.OperationReporter.WriteInformation(DesignStrings.RevertingSnapshot);
-            File.WriteAllText(modelSnapshotFile, modelSnapshotCode, Encoding.UTF8);
+
+            if (!dryRun)
+            {
+                File.WriteAllText(modelSnapshotFile, modelSnapshotCode, Encoding.UTF8);
+            }
         }
 
         return files;
@@ -375,8 +396,9 @@ public class MigrationsScaffolder : IMigrationsScaffolder
     /// <param name="projectDir">The project's root directory.</param>
     /// <param name="migration">The scaffolded migration.</param>
     /// <param name="outputDir">The directory to put files in. Paths are relative to the project directory.</param>
+    /// <param name="dryRun">If true, then nothing is actually written to disk.</param>
     /// <returns>The saved migrations files.</returns>
-    public virtual MigrationFiles Save(string projectDir, ScaffoldedMigration migration, string? outputDir)
+    public virtual MigrationFiles Save(string projectDir, ScaffoldedMigration migration, string? outputDir, bool dryRun)
     {
         var lastMigrationFileName = migration.PreviousMigrationId + migration.FileExtension;
         var migrationDirectory = outputDir ?? GetDirectory(projectDir, lastMigrationFileName, migration.MigrationSubNamespace);
@@ -387,19 +409,24 @@ public class MigrationsScaffolder : IMigrationsScaffolder
         var modelSnapshotFile = Path.Combine(modelSnapshotDirectory, modelSnapshotFileName);
 
         Dependencies.OperationReporter.WriteVerbose(DesignStrings.WritingMigration(migrationFile));
-        Directory.CreateDirectory(migrationDirectory);
-        File.WriteAllText(migrationFile, migration.MigrationCode, Encoding.UTF8);
-        File.WriteAllText(migrationMetadataFile, migration.MetadataCode, Encoding.UTF8);
 
-        Dependencies.OperationReporter.WriteVerbose(DesignStrings.WritingSnapshot(modelSnapshotFile));
-        Directory.CreateDirectory(modelSnapshotDirectory);
-        File.WriteAllText(modelSnapshotFile, migration.SnapshotCode, Encoding.UTF8);
+        if (!dryRun)
+        {
+            Directory.CreateDirectory(migrationDirectory);
+            File.WriteAllText(migrationFile, migration.MigrationCode, Encoding.UTF8);
+            File.WriteAllText(migrationMetadataFile, migration.MetadataCode, Encoding.UTF8);
+
+            Dependencies.OperationReporter.WriteVerbose(DesignStrings.WritingSnapshot(modelSnapshotFile));
+            Directory.CreateDirectory(modelSnapshotDirectory);
+            File.WriteAllText(modelSnapshotFile, migration.SnapshotCode, Encoding.UTF8);
+        }
 
         return new MigrationFiles
         {
             MigrationFile = migrationFile,
             MetadataFile = migrationMetadataFile,
-            SnapshotFile = modelSnapshotFile
+            SnapshotFile = modelSnapshotFile,
+            Migration = migration
         };
     }
 
@@ -464,7 +491,9 @@ public class MigrationsScaffolder : IMigrationsScaffolder
     /// <param name="fileName">The filename.</param>
     /// <returns>The file path or null if none.</returns>
     protected virtual string? TryGetProjectFile(string projectDir, string fileName)
-        => Directory.EnumerateFiles(projectDir, fileName, SearchOption.AllDirectories).FirstOrDefault();
+        => Directory.Exists(projectDir)
+            ? Directory.EnumerateFiles(projectDir, fileName, SearchOption.AllDirectories).FirstOrDefault()
+            : null;
 
     private bool ContainsForeignMigrations(string migrationsNamespace)
         => (from t in Dependencies.MigrationsAssembly.Assembly.GetConstructibleTypes()
